requirements.txt
appdirs==1.4.4
asgiref==3.3.1
distlib==0.3.1
Django==3.1.7
django-debug-toolbar==3.2
django-phonenumber-field==5.0.0
filelock==3.0.12
isort==5.8.0
packaging==20.9
phonenumbers==8.12.20
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tox==3.23.0
virtualenv==20.4.3
